backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-socketio==5.11.1
celery[redis]==5.3.6
redis==5.0.1
httpx==0.26.0
anthropic==0.18.1
telethon==1.34.0
qdrant-client==1.7.3
sentence-transformers==2.5.1
twilio==8.13.0
cryptography==42.0.4
shapely==2.0.3
geopy==2.4.1
